+++ pkg/integration/components/assert.go
@@ -1,4 +1,4 @@
-package helpers
+package components
 
 import (
 	"fmt"
