+++ pkg/integration/components/input.go
@@ -1,4 +1,4 @@
-package helpers
+package components
 
 import (
 	"fmt"
