+++ pkg/integration/components/shell.go
@@ -1,4 +1,4 @@
-package helpers
+package components
 
 import (
 	"fmt"
