+++ pkg/integration/tests/branch/suggestions.go
@@ -2,15 +2,15 @@ package branch
 
 import (
 	"github.com/jesseduffield/lazygit/pkg/config"
-	"github.com/jesseduffield/lazygit/pkg/integration/helpers"
+	"github.com/jesseduffield/lazygit/pkg/integration/components"
 )
 
-var Suggestions = helpers.NewIntegrationTest(helpers.NewIntegrationTestArgs{
+var Suggestions = components.NewIntegrationTest(components.NewIntegrationTestArgs{
 	Description:  "Checking out a branch with name suggestions",
 	ExtraCmdArgs: "",
 	Skip:         false,
 	SetupConfig:  func(config *config.AppConfig) {},
-	SetupRepo: func(shell *helpers.Shell) {
+	SetupRepo: func(shell *components.Shell) {
 		shell.
 			EmptyCommit("my commit message").
 			NewBranch("new-branch").
@@ -20,7 +20,7 @@ var Suggestions = helpers.NewIntegrationTest(helpers.NewIntegrationTestArgs{
 			NewBranch("other-new-branch-2").
 			NewBranch("other-new-branch-3")
 	},
-	Run: func(shell *helpers.Shell, input *helpers.Input, assert *helpers.Assert, keys config.KeybindingConfig) {
+	Run: func(shell *components.Shell, input *components.Input, assert *components.Assert, keys config.KeybindingConfig) {
 		input.SwitchToBranchesWindow()
 
 		input.PressKeys(keys.Branches.CheckoutBranchByName)
