+++ pkg/integration/tests/commit/commit.go
@@ -2,19 +2,19 @@ package commit
 
 import (
 	"github.com/jesseduffield/lazygit/pkg/config"
-	"github.com/jesseduffield/lazygit/pkg/integration/helpers"
+	"github.com/jesseduffield/lazygit/pkg/integration/components"
 )
 
-var Commit = helpers.NewIntegrationTest(helpers.NewIntegrationTestArgs{
+var Commit = components.NewIntegrationTest(components.NewIntegrationTestArgs{
 	Description:  "Staging a couple files and committing",
 	ExtraCmdArgs: "",
 	Skip:         false,
 	SetupConfig:  func(config *config.AppConfig) {},
-	SetupRepo: func(shell *helpers.Shell) {
+	SetupRepo: func(shell *components.Shell) {
 		shell.CreateFile("myfile", "myfile content")
 		shell.CreateFile("myfile2", "myfile2 content")
 	},
-	Run: func(shell *helpers.Shell, input *helpers.Input, assert *helpers.Assert, keys config.KeybindingConfig) {
+	Run: func(shell *components.Shell, input *components.Input, assert *components.Assert, keys config.KeybindingConfig) {
 		assert.CommitCount(0)
 
 		input.Select()
