+++ pkg/integration/tests/commit/new_branch.go
@@ -2,21 +2,21 @@ package commit
 
 import (
 	"github.com/jesseduffield/lazygit/pkg/config"
-	"github.com/jesseduffield/lazygit/pkg/integration/helpers"
+	"github.com/jesseduffield/lazygit/pkg/integration/components"
 )
 
-var NewBranch = helpers.NewIntegrationTest(helpers.NewIntegrationTestArgs{
+var NewBranch = components.NewIntegrationTest(components.NewIntegrationTestArgs{
 	Description:  "Creating a new branch from a commit",
 	ExtraCmdArgs: "",
 	Skip:         false,
 	SetupConfig:  func(config *config.AppConfig) {},
-	SetupRepo: func(shell *helpers.Shell) {
+	SetupRepo: func(shell *components.Shell) {
 		shell.
 			EmptyCommit("commit 1").
 			EmptyCommit("commit 2").
 			EmptyCommit("commit 3")
 	},
-	Run: func(shell *helpers.Shell, input *helpers.Input, assert *helpers.Assert, keys config.KeybindingConfig) {
+	Run: func(shell *components.Shell, input *components.Input, assert *components.Assert, keys config.KeybindingConfig) {
 		assert.CommitCount(3)
 
 		input.SwitchToCommitsWindow()
